build(aio): move attribute utils to helpers folder
This allows these utility functions to be reused across packages.
